Refactor mxutil
