+++ LEGO1/omni/src/action/mxdsaction.cpp
@@ -2,7 +2,7 @@
 
 #include "mxmisc.h"
 #include "mxtimer.h"
-#include "mxutil.h"
+#include "mxutilities.h"
 
 #include <float.h>
 #include <limits.h>
