+++ LEGO1/omni/src/audio/mxwavepresenter.cpp
@@ -7,7 +7,7 @@
 #include "mxmisc.h"
 #include "mxomni.h"
 #include "mxsoundmanager.h"
-#include "mxutil.h"
+#include "mxutilities.h"
 
 DECOMP_SIZE_ASSERT(MxWavePresenter, 0x6c);
 DECOMP_SIZE_ASSERT(MxWavePresenter::WaveFormat, 0x18);
