+++ LEGO1/omni/src/common/mxutilities.cpp
@@ -1,4 +1,4 @@
-#include "mxutil.h"
+#include "mxutilities.h"
 
 #include "mxcompositepresenter.h"
 #include "mxdsaction.h"
