+++ LEGO1/omni/src/stream/mxstreamcontroller.cpp
@@ -8,7 +8,7 @@
 #include "mxstl/stlcompat.h"
 #include "mxstreamchunk.h"
 #include "mxtimer.h"
-#include "mxutil.h"
+#include "mxutilities.h"
 
 DECOMP_SIZE_ASSERT(MxStreamController, 0x64)
 DECOMP_SIZE_ASSERT(MxNextActionDataStart, 0x14)
