+++ LEGO1/omni/src/video/mxdisplaysurface.cpp
@@ -2,7 +2,7 @@
 
 #include "mxmisc.h"
 #include "mxomni.h"
-#include "mxutil.h"
+#include "mxutilities.h"
 #include "mxvideomanager.h"
 
 #include <windows.h>
